chore: edit table row tests revamp
