+++ src/tests/components/EventsTableRow.test.tsx
@@ -1,10 +1,12 @@
 import { getAllByRole, render, screen } from "@testing-library/react";
 import EventsTableRow from "../../components/EventsTableRow";
 import { db } from "../mocks/db";
-import { EventItem } from "../../types";
-import { convertDateForTable } from "../../helpers/date";
+import { EventInput, EventItem } from "../../types";
+import { convertDateForInput, convertDateForTable } from "../../helpers/date";
 import userEvent from "@testing-library/user-event";
 import { SelectedItem } from "../../components/EventsTable";
+import { NullableEventInput, validFormData } from "../mocks/data";
+import AllProviders from "../AllProviders";
 
 describe("EventsTableRow", () => {
   let product: EventItem;
@@ -17,7 +19,7 @@ describe("EventsTableRow", () => {
     db.event.delete({ where: { id: { equals: product.id } } });
   });
 
-  const renderComponent = () => {
+  const renderComponent = (props?: SelectedItem | null) => {
     const mockedSetSelectedItem = vi.fn();
     const productIndex = 0;
 
@@ -25,9 +27,10 @@ describe("EventsTableRow", () => {
       <EventsTableRow
         idx={productIndex}
         item={product}
-        selectedItem={null}
+        selectedItem={props || null}
         setSelectedItem={mockedSetSelectedItem}
-      />
+      />,
+      { wrapper: AllProviders }
     );
 
     return {
@@ -35,6 +38,11 @@ describe("EventsTableRow", () => {
       user: userEvent.setup(),
       getEditButton: () => screen.queryByRole("button", { name: /edit/i }),
       getDeleteButton: () => screen.queryByRole("button", { name: /delete/i }),
+      getSaveButton: () => screen.queryByRole("button", { name: /save/i }),
+      getCloseButton: () => screen.queryByRole("button", { name: /close/i }),
+      getConfirmButton: () =>
+        screen.queryByRole("button", { name: /confirm/i }),
+      getCancelButton: () => screen.queryByRole("button", { name: /cancel/i }),
       rerender,
       mockedSetSelectedItem,
     };
@@ -86,6 +94,10 @@ describe("EventsTableRow", () => {
         rerender,
         getEditButton,
         getDeleteButton,
+        getSaveButton,
+        getCancelButton,
+        getConfirmButton,
+        getCloseButton,
         mockedSetSelectedItem,
       } = renderComponent();
 
@@ -113,21 +125,158 @@ describe("EventsTableRow", () => {
 
       if (actionType === "edit") {
         expect(getEditButton()).not.toBeInTheDocument();
-        expect(
-          screen.queryByRole("button", { name: /save/i })
-        ).toBeInTheDocument();
-        expect(
-          screen.queryByRole("button", { name: /close/i })
-        ).toBeInTheDocument();
+        expect(getSaveButton()).toBeInTheDocument();
+        expect(getCloseButton()).toBeInTheDocument();
       } else {
         expect(getDeleteButton()).not.toBeInTheDocument();
-        expect(
-          screen.queryByRole("button", { name: /confirm/i })
-        ).toBeInTheDocument();
-        expect(
-          screen.queryByRole("button", { name: /cancel/i })
-        ).toBeInTheDocument();
+        expect(getConfirmButton()).toBeInTheDocument();
+        expect(getCancelButton()).toBeInTheDocument();
       }
     }
   );
+
+  describe("EditTableRow", () => {
+    const user = userEvent.setup();
+
+    const getEventForm = async () => {
+      const nameInput = screen.getByPlaceholderText(/name/i);
+      const locationInput = screen.getByPlaceholderText(/location/i);
+      const entranceFeeInput = screen.getByPlaceholderText(/fee/i);
+      const organiserInput = screen.getByPlaceholderText(/organiser/i);
+      const dateInput = screen.getByLabelText(/date/i);
+
+      const fillForm = async (formData: NullableEventInput<EventInput>) => {
+        if (formData?.name) {
+          await user.type(nameInput, formData.name);
+        } else {
+          await user.clear(nameInput);
+        }
+
+        if (formData?.location) {
+          await user.type(locationInput, formData.location);
+        } else {
+          await user.clear(locationInput);
+        }
+
+        if (formData?.entranceFee) {
+          await user.type(entranceFeeInput, formData.entranceFee.toString());
+        } else {
+          await user.clear(entranceFeeInput);
+        }
+
+        if (formData?.organiser) {
+          await user.type(organiserInput, formData.organiser);
+        } else {
+          await user.clear(organiserInput);
+        }
+
+        if (formData?.eventDate) {
+          await user.type(dateInput, formData.eventDate);
+        } else {
+          await user.clear(dateInput);
+        }
+      };
+
+      return {
+        nameInput,
+        locationInput,
+        entranceFeeInput,
+        organiserInput,
+        dateInput,
+        fillForm,
+      };
+    };
+
+    it("Should render form filled with data ", async () => {
+      const { getSaveButton, getCloseButton } = renderComponent({
+        actionType: "edit",
+        id: product.id!,
+      });
+
+      const {
+        nameInput,
+        dateInput,
+        organiserInput,
+        entranceFeeInput,
+        locationInput,
+      } = await getEventForm();
+
+      expect(nameInput).toHaveValue(product.name);
+      expect(dateInput).toHaveValue(convertDateForInput(product.eventDate));
+      expect(organiserInput).toHaveValue(product.organiser);
+      expect(entranceFeeInput).toHaveValue(product.entranceFee);
+      expect(locationInput).toHaveValue(product.location);
+
+      expect(getSaveButton()).toBeInTheDocument();
+      expect(getCloseButton()).toBeInTheDocument();
+    });
+
+    it("Should disable the save button when form is not valid", async () => {
+      const { getSaveButton } = renderComponent({
+        actionType: "edit",
+        id: product.id!,
+      });
+
+      const { fillForm } = await getEventForm();
+
+      await fillForm({ ...validFormData, name: "" });
+      expect(getSaveButton()).toBeDisabled();
+    });
+
+    it("Should submit form when 'save' button is clicked, and close the row", async () => {
+      const {
+        getEditButton,
+        getSaveButton,
+        getCloseButton,
+        rerender,
+        productIndex,
+        user,
+        mockedSetSelectedItem,
+      } = renderComponent({
+        actionType: "edit",
+        id: product.id!,
+      });
+
+      await user.click(getSaveButton()!);
+
+      rerender(
+        <EventsTableRow
+          idx={productIndex}
+          item={product}
+          selectedItem={null}
+          setSelectedItem={mockedSetSelectedItem}
+        />
+      );
+
+      expect(mockedSetSelectedItem).toHaveBeenCalledWith(null);
+      expect(getSaveButton()).not.toBeInTheDocument();
+      expect(getCloseButton()).not.toBeInTheDocument();
+      expect(getEditButton()).toBeInTheDocument();
+    });
+
+    it("Should close the row when 'close' button is clicked", async () => {
+      const {
+        getCloseButton,
+        getEditButton,
+        productIndex,
+        rerender,
+        user,
+        mockedSetSelectedItem,
+      } = renderComponent({ actionType: "edit", id: product.id! });
+
+      await user.click(getCloseButton()!);
+
+      rerender(
+        <EventsTableRow
+          idx={productIndex}
+          item={product}
+          selectedItem={null}
+          setSelectedItem={mockedSetSelectedItem}
+        />
+      );
+
+      expect(mockedSetSelectedItem).toHaveBeenCalledWith(null);
+      expect(getEditButton()).toBeInTheDocument();
+    });
+  });
 });
